Switch to text/html for paste where we can.
Build special URLs to detect the same host being in-use, and much more.

+++ wsd/LOOLWSD.hpp
@@ -45,7 +45,7 @@ public:
 
     // An Application is a singleton anyway,
     // so just keep these as statics.
-    static std::atomic<unsigned> NextSessionId;
+    static std::atomic<uint64_t> NextSessionId;
     static unsigned int NumPreSpawnedChildren;
     static bool NoCapsForKit;
     static bool NoSeccomp;
